refactor(bikes): group services and constrain admin status updates
