+++ apps/server/src/domain/bikes/index.ts
@@ -4,7 +4,8 @@ export * from "./repository/bike-command.repository";
 export * from "./repository/bike-query.repository";
 export * from "./repository/bike-stats.repository";
 export * from "./repository/bike.repository";
-export * from "./services/bike-stats.service";
-export * from "./services/bike.service";
-export * from "./services/bike.service.types";
-export * from "./services/soft-delete-bike.service";
+export * from "./services/commands/bike-command.service";
+export * from "./services/commands/bike-command.service.types";
+export * from "./services/commands/soft-delete-bike.service";
+export * from "./services/queries/bike-stats.service";
+export * from "./services/queries/bike-stats.service.types";
