+++ apps/server/src/domain/bikes/services/commands/bike-command.service.types.ts
@@ -8,23 +8,36 @@ import type {
   BikeCurrentlyReserved,
   BikeNotFound,
   BikeRepositoryError,
-  BikeStationPlacementCapacityExceeded,
   BikeStationNotFound,
+  BikeStationPlacementCapacityExceeded,
   BikeSupplierNotFound,
   InvalidBikeStatus,
-} from "../domain-errors";
-import type { BikeFilter, BikeRow, BikeSortField } from "../models";
-import type { BikeUpdatePatch } from "../repository/bike.repository.types";
+} from "../../domain-errors";
+import type { BikeFilter, BikeRow, BikeSortField } from "../../models";
 
 export type BikeManageableStatus = Extract<BikeStatus, "AVAILABLE" | "BROKEN">;
+export type AdminBikeManageableStatus = Extract<BikeStatus, "AVAILABLE" | "BROKEN" | "MAINTAINED" | "UNAVAILABLE">;
+
+export type CreateBikeInput = {
+  stationId: string;
+  supplierId: string;
+  status?: BikeStatus;
+};
+
+export type BikeStationScopedStatusUpdateInput = {
+  stationId: string;
+  status: BikeManageableStatus;
+};
+
+export type AdminBikeUpdatePatch = Partial<{
+  stationId: string;
+  status: AdminBikeManageableStatus;
+  supplierId: string | null;
+}>;
 
 export type BikeService = {
   createBike: (
-    input: {
-      stationId: string;
-      supplierId: string;
-      status?: BikeStatus;
-    },
+    input: CreateBikeInput,
   ) => Effect.Effect<
     BikeRow,
     BikeRepositoryError | BikeStationNotFound | BikeStationPlacementCapacityExceeded | BikeSupplierNotFound
@@ -39,13 +52,14 @@ export type BikeService = {
 
   adminUpdateBike: (
     bikeId: string,
-    patch: BikeUpdatePatch,
+    patch: AdminBikeUpdatePatch,
   ) => Effect.Effect<
     Option.Option<BikeRow>,
     | BikeCurrentlyRented
     | BikeCurrentlyReserved
     | BikeNotFound
     | BikeRepositoryError
+    | InvalidBikeStatus
     | BikeStationPlacementCapacityExceeded
     | BikeStationNotFound
     | BikeSupplierNotFound
@@ -53,10 +67,7 @@ export type BikeService = {
 
   updateBikeStatusInStationScope: (
     bikeId: string,
-    input: {
-      stationId: string;
-      status: BikeManageableStatus;
-    },
+    input: BikeStationScopedStatusUpdateInput,
   ) => Effect.Effect<
     BikeRow,
     | BikeNotFound
